Actualizacion de API

+++ ApiBiblioteca/Controllers/ReaderController.cs
@@ -104,12 +104,18 @@ namespace ApiBiblioteca.Controllers
             try
             {
                 var readerToDelete = await readerRepository.GetReader(id);
+                var transactionsByReader = await readerRepository.GetTransactionsByReaderID(id);
 
                 if (readerToDelete == null)
                 {
                     return NotFound($"Reader with Id = {id} not found");
                 }
-                
+                if(transactionsByReader != 0)
+                {
+                    return StatusCode(StatusCodes.Status405MethodNotAllowed,
+                    $"Error deleting new reader record: Transactions found");
+                }
+
                 return await readerRepository.DeleteReader(id);
             }
             catch (Exception e)
